ajout controller et correction diverses

+++ microservice-paiement/microservice-paiement/src/main/java/com/mpaiement/repository/IPaiementRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.mpaiement.entity.Paiement;
 
 @Repository
-public interface IPaiementRepository extends JpaRepository<Paiement, Integer> {
+public interface IPaiementRepository extends JpaRepository<Paiement, Long> {
 
 }
