+++ microservice-reservation/microservice-reservation/src/main/java/com/mreservation/repository/IReservationRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.mreservation.entity.Reservation;
 
 @Repository
-public interface IReservationRepository extends JpaRepository<Reservation, Integer>{
+public interface IReservationRepository extends JpaRepository<Reservation, Long>{
 
 	
 }
